move some shared methods to abstract class
